MediaPlayerPrivateGStreamerBase: use new render-rectangle property on waylandsink
To propagate the render-rectangle for the hole punch on the video, wayland-sink
has been updated to support render-rectangle property using GstOverlay interface.
This correctly sets the windows width and height.

As the surfaces aren't currently linked between Thunder and waylandsink the
compositor ignores the x,y part.

Previously we used the window-width and window-height properties found
in NXP waylandsink, but most other video sinks use render-rectangle.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>